chore(Table): Convert Table to TypeScript
