--- packages/forma-36-react-components/src/components/Table/TableBody/TableBody.js
@@ -1,21 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-
-export class TableBody extends React.Component {
-  static propTypes = {
-    extraClassNames: PropTypes.string,
-    children: PropTypes.node.isRequired,
-  };
-
-  static defaultProps = {
-    extraClassNames: '',
-  };
-
-  render() {
-    const { extraClassNames, children } = this.props;
-
-    return <tbody className={extraClassNames}>{children}</tbody>;
-  }
-}
-
-export default TableBody;
