+++ packages/forma-36-react-components/src/components/Table/TableBody/TableBody.tsx
@@ -1,0 +1,16 @@
+import React, { Component } from 'react';
+
+export interface TableBodyProps {
+  extraClassNames?: string;
+  children: React.ReactNode;
+}
+
+export class TableBody extends Component<TableBodyProps> {
+  render() {
+    const { extraClassNames, children } = this.props;
+
+    return <tbody className={extraClassNames}>{children}</tbody>;
+  }
+}
+
+export default TableBody;
