+++ packages/forma-36-react-components/src/components/Table/TableCell/TableCell/TableCellContext.ts
@@ -1,8 +1,8 @@
 import React from 'react';
 
 export const contextOptions = {
-  body: { name: 'body', element: 'td' },
-  head: { name: 'head', element: 'th' },
+  body: { name: 'body', element: 'td', offsetTop: 0 },
+  head: { name: 'head', element: 'th', offsetTop: 0 },
 };
 
 export const TableCellContext = React.createContext(contextOptions.body);
