+++ packages/forma-36-react-components/src/components/Table/TableCell/TableSortingLabel/index.ts
