+++ packages/forma-36-react-components/src/components/Table/TableHead/index.ts
